needs norminette testing

+++ ft_printf/ft_printf.h
@@ -26,7 +26,7 @@ int	ft_put_c(char c);
 int	ft_put_s(char *s);
 int	ft_put_id(int  n);
 int	ft_put_u(unsigned int n);
-int     ft_put_xX(unsigned int n, char format);
+int ft_put_xX(unsigned long long n, char format);
 int	ft_put_p(unsigned long long p);
 
 #endif
